restructure API folders in fastspring
