+++ tests/api/fastspring/fastspring_order_refund_scenario_test.js
@@ -1,12 +1,12 @@
 var faker = require('faker');
-const payload = require('./payload');
-Feature('Order create and partialrefund');
+const payload = require('./fastspring_payload');
+Feature('Order create and refund');
 
-Scenario('partial refund', async({ I }) => {
+Scenario('Single product order and Full-Refund Test', async({ I }) => {
     for (let i = 1; i <= 2; i++) {
         const data = payload.getFakerData();
         const data_order = payload.getFastSpringOrder(data);
-        const data_ref = payload.getPartialReturn(data);
+        const data_ref = payload.getFullRefund(data);
         if (i % 2 == 0) {
             await I.sendPostRequest('/webhook/fastspring/7d494986-fe98-413e-b671-565a8dfaae30', data_order).then((res) => {
                 I.assertEqual(res.status, 200);
@@ -24,5 +24,6 @@ Scenario('partial refund', async({ I }) => {
             });
         }
     }
+
 
 });
