+++ tests/api/fastspring/fastspring_single_refund_test.js
@@ -1,5 +1,5 @@
 Feature('Fastspring Single Orderrefund');
-const payload = require('./payload');
+const payload = require('./fastspring_payload');
 Scenario('Fastspring Single order Placeref', async({ I }) => {
     var load = payload.getRefund();
     const res = await I.sendPostRequest('/webhook/fastspring/7d494986-fe98-413e-b671-565a8dfaae30', load);
