+++ tests/api/fastspring/fastspring_subs_order_refund_scenario_test.js
@@ -1,12 +1,12 @@
 var faker = require('faker');
-const payload = require('./payload');
-Feature('Order create and refund');
+const payload = require('./fastspring_payload');
+Feature('Subscription order & refund');
 
-Scenario('Add & Deactivate Track', async({ I }) => {
+Scenario('Subscription type product order and return', async({ I }) => {
     for (let i = 1; i <= 2; i++) {
         const data = payload.getFakerData();
-        const data_order = payload.getFastSpringOrder(data);
-        const data_ref = payload.getRefund(data);
+        const data_order = payload.getFastspringSubsorder(data);
+        const data_ref = payload.getFastspringSubreturn(data);
         if (i % 2 == 0) {
             await I.sendPostRequest('/webhook/fastspring/7d494986-fe98-413e-b671-565a8dfaae30', data_order).then((res) => {
                 I.assertEqual(res.status, 200);
