+++ tests/api/tracking_scenario_test.js
@@ -1,9 +1,9 @@
 var faker = require('faker');
 const payload = require('./payload');
-Feature('Site activate and Deactivate Scenario');
+Feature('Appsero Analytics');
 
-Scenario('Add & Deactivate Track', async({ I }) => {
-    for (let i = 1; i <= 10; i++) {
+Scenario('Site activate and Deactivate Scenario', async({ I }) => {
+    for (let i = 1; i <= 2; i++) {
         const fake_data = payload.getFakerData();
         const getrack = payload.getTrackingInfo(fake_data);
         if (i % 2 == 0) {
